backend/requirements.txt
flask
flask-cors
requests
python-dotenv
openai
gradio_client
pydub